+++ R/utils.R
@@ -8,10 +8,19 @@
   x
 }
 
-# define a list of utilities to be made available when rendering
-render_time_utils <- list(
+# list of utilities to be made available when filling templates / rendering
+fill_render_utils <- list(
   `%||%` = `%||%`
 )
+
+# environment with fill- / render-time utilities, including `site_path()` for
+# constructing paths relative to side_dir
+fill_render_env <- function(site_dir, parent = parent.frame()) {
+  env <- list2env(fill_render_utils, parent = parent)
+  env$site_path <- function(...) file.path(site_dir, ...)
+  env
+}
+
 
 toQuotedString <- function(x) {
   toString(sQuote(x, q = FALSE))